Fix LuaJ crashing the compiler
LuaJ is, as we both know, awful. This is a particular case of this
awfulness where what can only be presumed to be an unsupported pattern
causes execution to fail with an unhelpful error message. Thankfully the
documentation handling code is the only place where we use such
complicated patterns, and slapping a `handler-case` around it wasn't too
hard.

Note that we do not (and *can not*) promise compatibility with
non-conforming Lua implementations: I'm going out of my way to make this
work because someone actually reported it. Unfortunately my computer is
not beefy enough (and I presume no computer is) to run the test suite
under LuaJ.